# docs/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Portfolio-ready authentication API built with Gin, MongoDB and JWT.",
    "title": "Go Auth API",
    "contact": {
      "name": "Kaio",
      "url": "https://github.com/"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.1.0"
  },
  "host": "127.0.0.1:8080",
  "basePath": "/",
  "paths": {
    "/": {
      "get": {
        "description": "Returns metadata about the API and the most important endpoints.",
        "produces": [
          "application/json"
        ],
        "summary": "API overview",
        "responses": {
          "200": {
            "description": "API metadata",
            "schema": {
              "$ref": "#/definitions/dtos.APIInfoResponse"
            }
          }
        }
      }
    },
    "/health": {
      "get": {
        "description": "Checks whether the API is up and returns its current version.",
        "produces": [
          "application/json"
        ],
        "summary": "Health check",
        "responses": {
          "200": {
            "description": "Service is healthy",
            "schema": {
              "$ref": "#/definitions/dtos.HealthResponse"
            }
          }
        }
      }
    },
    "/api/v1/users": {
      "post": {
        "description": "Registers a new user account in the API.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "users"
        ],
        "summary": "Create a new user",
        "parameters": [
          {
            "description": "User payload",
            "name": "user",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dtos.CreateUserRequest"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully",
            "schema": {
              "$ref": "#/definitions/dtos.Message"
            }
          },
          "400": {
            "description": "Validation error",
            "schema": {
              "$ref": "#/definitions/dtos.APIError"
            }
          },
          "409": {
            "description": "User already exists",
            "schema": {
              "$ref": "#/definitions/dtos.APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/dtos.APIError"
            }
          }
        }
      }
    },
    "/api/v1/users/login": {
      "post": {
        "description": "Validates credentials and returns a JWT token.",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "tags": [
          "auth"
        ],
        "summary": "Login",
        "parameters": [
          {
            "description": "Login payload",
            "name": "credentials",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/dtos.LoginUserRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Authenticated successfully",
            "schema": {
              "$ref": "#/definitions/dtos.LoginUserResponse"
            }
          },
          "400": {
            "description": "Invalid credentials or request body",
            "schema": {
              "$ref": "#/definitions/dtos.APIError"
            }
          },
          "500": {
            "description": "Internal server error",
            "schema": {
              "$ref": "#/definitions/dtos.APIError"
            }
          }
        }
      }
    }
  },
  "definitions": {
    "dtos.APIError": {
      "type": "object",
      "properties": {
        "error": {
          "type": "string"
        }
      }
    },
    "dtos.APIInfoResponse": {
      "type": "object",
      "properties": {
        "description": {
          "type": "string",
          "example": "JWT authentication API built with Gin and MongoDB."
        },
        "docsUrl": {
          "type": "string",
          "example": "/swagger/index.html"
        },
        "endpoints": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "name": {
          "type": "string",
          "example": "Go Auth API"
        },
        "version": {
          "type": "string",
          "example": "1.1.0"
        }
      }
    },
    "dtos.CreateUserRequest": {
      "type": "object",
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string",
          "example": "junior.dev@email.com"
        },
        "password": {
          "type": "string",
          "minLength": 6,
          "example": "123456"
        }
      }
    },
    "dtos.HealthResponse": {
      "type": "object",
      "properties": {
        "service": {
          "type": "string",
          "example": "Go Auth API"
        },
        "status": {
          "type": "string",
          "example": "ok"
        },
        "timestamp": {
          "type": "string",
          "example": "2026-04-17T14:00:00Z"
        },
        "version": {
          "type": "string",
          "example": "1.1.0"
        }
      }
    },
    "dtos.LoginUserRequest": {
      "type": "object",
      "required": [
        "email",
        "password"
      ],
      "properties": {
        "email": {
          "type": "string",
          "example": "junior.dev@email.com"
        },
        "password": {
          "type": "string",
          "example": "123456"
        }
      }
    },
    "dtos.LoginUserResponse": {
      "type": "object",
      "properties": {
        "success": {
          "type": "boolean",
          "example": true
        },
        "token": {
          "type": "string",
          "example": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        }
      }
    },
    "dtos.Message": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string"
        }
      }
    }
  },
  "securityDefinitions": {
    "BearerAuth": {
      "description": "Use the format: Bearer <token>",
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}